+++ libs/PF_preview.js
@@ -103,7 +103,7 @@ const resultReceivedHandler = ( result ) => {
 				'border': 'none',
 				'overflow': 'hidden'
 			} )
-			.load( loadFrameHandler )
+			.on( 'load', loadFrameHandler )
 			.appendTo( $previewpane );
 
 		}